use HashMap indexed by users instead of adding all messages to the same message queue
